[Add] List 클릭 이벤트 변수 넘기기

+++ wearos/app/src/main/java/com/example/kkobak/room/adapter/TodoAdapter.java
@@ -50,7 +50,7 @@ public class TodoAdapter extends RecyclerView.Adapter<TodoAdapter.ViewHolder>{
     public class ViewHolder extends RecyclerView.ViewHolder{
 
         TextView textView;
-        int chlId;
+        Long chlId;
         boolean done;
 
         public ViewHolder(View itemView){
